Add handling for pre tags
Previously, <pre> tags in an epub were falling through to the default
case and just being rendered as text, which was stringing all the
individual lines together into a mess. This adds special handling for
pre tags by rendered each line of text individually, adding a single
line of whitespace above and below, and a tab character at the start of
each line. This can lead to broken text wrapping in a case where a <pre>
block contains a long line (though this is arguably the intended effect of
the tag).