Jump forward added.
This lets the jump stack be traversed as a double linked list
jump 1 <-> jump 2 <-> jump 3 <- Last jump

This is implemented by adding a int that points to the top of array.
stack_pointer
Array elements are always picked one less than the top.
The destination of the last jump destination is not in the stack. So jump forward can't really move to it.
But the jump n - 1 should point to the link that you press to make it.

When you click a link and stack_pointer is pointing to last
jump 1 <-> jump 2 <-> jump 3
                             *stack_pointer
jump 1 <-> jump 2 <-> jump 3 <-> jump 4
                                        *stack_pointer + 1

However, when clicking a link on a stack_pointer pointing before that:
jump 1 <-> jump 2 <-> jump 3 <-> jump 4
                   *stack_pointer

jump 1 <-> jump 2 <-> new jump 3
                                  *stack_pointer